Fix duplicate CI runs
GitHub runs both the `push` and `pull_request` workflows for PRs that
are coming from our own repo, which runs double the CI builds wasting
resources.

The downside of this change is that if we add multiple development lines
in the future, they'll need adding here to get CI checks.